examples/delta-dtm-line-plot/pyproject.toml
[project]
name = "delta-dtm-line-plot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.6",
    "azure-storage-blob>=12.24.1",
    "duckdb>=1.1.3",
    "geopandas>=1.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "pystac-client>=0.8.5",
    "rioxarray>=0.18.2",
    "xarray>=2024.10.0",
]
